requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic>=0.40.0
gspread==6.1.2
google-auth==2.30.0
httpx==0.28.1
python-dotenv==1.0.1